wrap the connection for retries and failures on initial opening
- new genserver to handle and catch the WS client connection errors
- moved the main functions in to Websocket
